requirements.txt
Flask==3.0.3
pandas==2.2.2
python-dotenv==1.0.1
gunicorn==21.2.0
